# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Smart Web Scraper API
  description: 지능형 웹 데이터 수집 플랫폼 — 5단계 적응형 파이프라인 + Self-Healing
  version: 1.0.0
  contact:
    name: Smart Web Scraper
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: 로그인 (JWT 발급)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        "200":
          description: 로그인 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  refresh_token: { type: string }
                  token_type: { type: string, example: bearer }
        "401":
          description: 인증 실패
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Access Token 재발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200": { description: 재발급 성공 }
        "401": { description: 유효하지 않은 refresh token }
...
  /api/v1/dashboard/success-rate:
    get:
      tags: [Dashboard]
      summary: 성공률 추이
      parameters:
        - name: days
          in: query
          schema: { type: integer, default: 30 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date: { type: string, format: date }
                    success_rate: { type: number }
...
  /api/v1/dashboard/alerts:
    get:
      tags: [Dashboard]
      summary: 알림 목록
      parameters:
        - name: is_read
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AlertResponse" }
  /api/v1/dashboard/alerts/{id}/read:
    put:
      tags: [Dashboard]
      summary: 알림 읽음 처리
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: 읽음 처리됨 }
...
  /api/v1/history:
    get:
      tags: [History]
      summary: 방문 이력 목록
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 20 }
        - name: url_id
          in: query
          schema: { type: string, format: uuid }
        - name: success
          in: query
          schema: { type: boolean }
        - name: pipeline_name
          in: query
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/VisitLogResponse" }
                  total: { type: integer }
                  pages: { type: integer }
  /api/v1/history/export:
    get:
      tags: [History]
      summary: 방문 이력 내보내기 (CSV / JSON)
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [csv, json], default: csv }
        - name: url_id
          in: query
          schema: { type: string, format: uuid }
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: 파일 다운로드
          content:
            text/csv:
              schema: { type: string }
            application/json:
              schema: { type: array }
  /api/v1/history/{id}:
    get:
      tags: [History]
      summary: 방문 이력 상세
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VisitLogResponse" }
        "404": { description: 이력 없음 }

  # ─── WebSocket ────────────────────────────────────────────
...
components:
  schemas:
...
    VisitLogResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        url: { type: string }
        visited_at: { type: string, format: date-time }
        success: { type: boolean }
        pipeline_name: { type: string }
        duration_ms: { type: integer }
        status_code: { type: integer }
        error_type: { type: string }
        error_message: { type: string }
        pipelines_attempted: { type: array, items: { type: string } }
        content_size_bytes: { type: integer }
        items_extracted: { type: integer }
        content_hash: { type: string }
        antibot_detected: { type: string }
        captcha_encountered: { type: boolean }
        healing_applied: { type: boolean }
        healing_type: { type: string }
        method_details: { type: object }
    AlertResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        severity: { type: string, enum: [info, warning, error, critical] }
        alert_type: { type: string }
        message: { type: string }
        is_read: { type: boolean }
        created_at: { type: string, format: date-time }